+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+##
+
+### Added
+- NeXus writer function, phil template and parameter file for Electron Diffraction data.
+- CLI for Electron Diffraction.
+- Function to write a compressed copy of a dataset in a specified NXgroup. Main application: mask and flatfield in NXdetector.
+- General conversion to mcstas from a user defined coordinate frame.
+- NXcoordinate_system_set writer.
+- Tests for NXentry, NXcoordinate_system_set and NXnote writers.
+- Tests for Metafile tools and ElectronDiffraction tools.
+
+### Changed
+- Entry key for data in NXdata as argument.
+- Rotation axis for the detector not hard coded to two theta anymore.
+- Splitting of arrays done before feeding the goniometer/detector dictionaries to the NXclass writers.
+- All conversions to mcstas done before passing vectors/arrays to writers.
+
+### Fixed
+- Outstanding data_size order issues.
+
+---
 ## 0.6.11
 
 ### Added
@@ -18,7 +39,7 @@
 - Ability to specify non-zero indexed VDS.
 - New cli for I19-2 NeXus writer - it can now work with and without GDA-generated metadata.
 - Docstrings for namedtuples in beamline scripts.
-- Show phil parameters using `--show-config` from the command line. 
+- Show phil parameters using `--show-config` from the command line.
 
 ### Changed
 - SSX chip reading tools moved to SSX_chip.py